+++ app/src/main/java/com/techcourse/service/transaction/TransactionLogicExecutor.java
@@ -1,0 +1,7 @@
+package com.techcourse.service.transaction;
+
+@FunctionalInterface
+public interface TransactionLogicExecutor {
+
+    void run();
+}
