+++ app/src/main/java/com/techcourse/service/transaction/TransactionTemplate.java
@@ -1,7 +1,6 @@
-package com.techcourse.service;
+package com.techcourse.service.transaction;
 
 import com.techcourse.config.DataSourceConfig;
-import com.techcourse.domain.User;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.datasource.DataSourceUtils;
 import org.springframework.transaction.TransactionException;
@@ -11,31 +10,15 @@ import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.SQLException;
 
-public class TxUserService implements UserService {
+public class TransactionTemplate {
 
-    private final UserService userService;
-
-    public TxUserService(final UserService userService) {
-        this.userService = userService;
-    }
-
-    @Override
-    public User findById(final long id) {
-        return userService.findById(id);
-    }
-
-    @Override
-    public void insert(final User user) {
-        userService.insert(user);
-    }
-
-    @Override
-    public void changePassword(final long id, final String newPassword, final String createBy) {
+    public void executeWithTransaction(final TransactionLogicExecutor transactionLogicExecutor) {
         final DataSource dataSource = DataSourceConfig.getInstance();
         final Connection connection = DataSourceUtils.getConnection(dataSource);
         try {
             connection.setAutoCommit(false);
-            userService.changePassword(id, newPassword, createBy);
+
+            transactionLogicExecutor.run();
 
             connection.commit();
         } catch (final SQLException | DataAccessException e) {
